Create 1-2 Baseball Costs
Using Program 1-1 as an example, write a program that calculates how much a Little
League baseball team spent last year to purchase new baseballs. The program should
prompt the user to enter the number of baseballs purchased and the cost of each baseball.
It should then calculate and display the total amount spent to purchase the baseballs.